close implicitly create tagTranslator after terminal operation
Such as count(), sum(), …, reduce(), or stream().close(). This avoids dangling prepared statements to lie around after a query was performed. This is a regression/consequence of #224.

This also fixes the fact that in the current implementation, because of a typo/bug the tagInterpreter objects themselves were never persisted when calling `osmTag` methods. :-/

Note that this solution passes all tests, but doesn't work in the following simple example!

  @Test
  public void testFoo() throws Exception {
    MapReducer<OSMContribution> bar = createMapReducerOSMContribution()
        .timestamps(timestamps72)
        .osmEntityFilter(entity -> entity.getId() == 617308093)
        .osmTag("highway");
    assertTrue(bar.count() > -1);
    assertTrue(bar.osmTag("highway", "traffic_signals").count() > -1);
  }

Here, a runtime exception will be thrown, because when calling the first `bar.count()`, the internal tragTranslator of `bar` is closed and further tag translation requests cannot be answered anymore (as performed in the last line).